requirements.txt
aiogram>=3.6.0
python-dotenv>=1.0.0
google-auth>=2.25.2
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.111.0
aiosqlite>=0.19.0
redis>=5.0.1
pytz>=2023.3
